refactor(community): split post card composition from body
Extract the visual Column (header + guide lead + title + content +
media + actions) and the _GuideLeadBlock helper into a new stateless
CommunityPostCardBody. The composition root keeps handlers, lifecycle,
and the Card/InkWell wrapper; body takes callbacks and has no ref.
Card drops from 394 to ~170 lines. Smoke test added for the body.
